Refactor binaryqf impl
